LPS-52537 Article should be viewed if its in Pending status
This issue was caused by the status of an article.When the status of an article is Pending, it throws Exception while clicking it.
The root cause is that whenever the status is pending code below code tries to fetch data from database with status APPROVED :

KBArticle latestKBArticle = KBArticleLocalServiceUtil.getLatestKBArticle(kbArticle.getResourcePrimKey(), WorkflowConstants.STATUS_APPROVED);

So above code shouldn't be executed if status is pending, so as solution i have added !kbArticle.isPending() as validation, as it would not allow to execute above code if status is pending.

The method (which is causing this issue) is expecting a resourcePrimKey , but in this case (Pending status) getClassPK() returns getKbArticleId instead of resourcePrimKey, so here i made changes so that in case of pending status it would return resourcePrimKey.