LPS-53174 check if ZIP file is invalid
When given an invalid ZIP file, the reader `listFiles()` method will
return null. Catch the generated NPE and throw a more appropriate
exception.

A proper solution would be to change the way `ZipReader` behaves, but
that may introduce unknown compatibility issues.  A more adecuate
solution will be implemented in master.